+++ src/main/java/top/theillusivec4/curios/server/SlotHelper.java
@@ -28,18 +28,12 @@ import java.util.Set;
 import java.util.SortedMap;
 import java.util.TreeMap;
 import javax.annotation.Nonnull;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraftforge.network.PacketDistributor;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.type.ISlotType;
 import top.theillusivec4.curios.api.type.inventory.ICurioStacksHandler;
 import top.theillusivec4.curios.api.type.util.ISlotHelper;
 import top.theillusivec4.curios.common.inventory.CurioStacksHandler;
-import top.theillusivec4.curios.common.inventory.container.CuriosContainer;
-import top.theillusivec4.curios.common.network.NetworkHandler;
-import top.theillusivec4.curios.common.network.server.sync.SPacketSyncOperation;
-import top.theillusivec4.curios.common.network.server.sync.SPacketSyncOperation.Operation;
 
 public class SlotHelper implements ISlotHelper {
 
@@ -114,21 +108,9 @@ public class SlotHelper implements ISlotHelper {
 
   @Override
   public void growSlotType(String id, int amount, final LivingEntity livingEntity) {
-    CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity).ifPresent(handler -> {
-      handler.growSlotType(id, amount);
-
-      if (livingEntity instanceof ServerPlayer player) {
-        NetworkHandler.INSTANCE
-            .send(PacketDistributor.TRACKING_ENTITY_AND_SELF.with(() -> player),
-                new SPacketSyncOperation(livingEntity.getId(), id, Operation.GROW, amount));
-
-        if (player.containerMenu instanceof CuriosContainer) {
-          ((CuriosContainer) player.containerMenu).resetSlots();
-        }
-      }
-    });
+    CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity)
+        .ifPresent(handler -> handler.growSlotType(id, amount));
   }
-
 
   @Override
   public void shrinkSlotType(String id, final LivingEntity livingEntity) {
@@ -137,17 +119,28 @@ public class SlotHelper implements ISlotHelper {
 
   @Override
   public void shrinkSlotType(String id, int amount, final LivingEntity livingEntity) {
+    CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity)
+        .ifPresent(handler -> handler.shrinkSlotType(id, amount));
+  }
+
+  @Override
+  public void unlockSlotType(String id, LivingEntity livingEntity) {
     CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity).ifPresent(handler -> {
-      handler.shrinkSlotType(id, amount);
 
-      if (livingEntity instanceof ServerPlayer player) {
-        NetworkHandler.INSTANCE
-            .send(PacketDistributor.TRACKING_ENTITY_AND_SELF.with(() -> player),
-                new SPacketSyncOperation(livingEntity.getId(), id, Operation.SHRINK, amount));
+      if (handler.getStacksHandler(id).map(stacksHandler -> stacksHandler.getSlots() == 0)
+          .orElse(false)) {
+        handler.growSlotType(id, 1);
+      }
+    });
+  }
 
-        if (player.containerMenu instanceof CuriosContainer) {
-          ((CuriosContainer) player.containerMenu).resetSlots();
-        }
+  @Override
+  public void lockSlotType(String id, LivingEntity livingEntity) {
+    CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity).ifPresent(handler -> {
+      int amount = handler.getStacksHandler(id).map(ICurioStacksHandler::getSlots).orElse(0);
+
+      if (amount > 0) {
+        handler.shrinkSlotType(id, amount);
       }
     });
   }