+++ docs/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.18.1-5.0.5.2] - 2022.01.15
+### Fixed
+- Fixed deprecated usages of `ISlotHelper#unlockSlotType` and `ISlotHelper#lockSlotType` modifying slots differently
+from previous behavior
+- Fixed deprecated usages of `ISlotHelper#growSlotType` and `ISlotHelper#shrinkSlotType` logging client-side errors to
+the console while in the Curios inventory
+
 ## [1.18.1-5.0.5.1] - 2021.01.09
 ### Fixed
 - Fixed dragged items desyncing when switching between Curios and the Creative inventory screen [#202](https://github.com/TheIllusiveC4/Curios/issues/202)